02/03/2023

+++ AURORE RH/FRONDEND/AURORERH/AURORERH/src/app/Pages/listing-employer/listing-employer.component.ts
@@ -1,7 +1,11 @@
 import { Component, OnInit } from '@angular/core';
-import { DELETE_EMPLOYER, LIST_EMPLOYERS } from 'src/app/shared/_elements/api_constante';
+import { ActivatedRoute, Router } from '@angular/router';
+import { DELETE_EMPLOYER, LIST_EMPLOYERS, READBYID_EMPLOYER, UPDATE_EMPLOYER } from 'src/app/shared/_elements/api_constante';
 import { EmployerReponseModel } from 'src/app/shared/_models/responses/employer-response.model';
 import { EmployerService } from 'src/app/shared/_services/employerService';
+import { NotificationService } from 'src/app/shared/_services/notification.service';
+import Swal from 'sweetalert2'
+import { AjoutEmployerComponent } from '../ajout-employer/ajout-employer.component';
 
 @Component({
   selector: 'app-listing-employer',
@@ -10,17 +14,22 @@ import { EmployerService } from 'src/app/shared/_services/employerService';
 })
 export class ListingEmployerComponent implements OnInit {
 
-
+  public isLoading!: boolean;
   public employers: EmployerReponseModel[] = [];
+  public id!: any;
+
   constructor(
     private employerService: EmployerService,
+    private notif: NotificationService,
+    private router: Router,
+    private route: ActivatedRoute,
   ) { }
 
   ngOnInit(): void {
     this.getEmployer();
 
-
   }
+
   getEmployer(){
     this.employerService.get(LIST_EMPLOYERS).then((response:any)=>{
       this.employers = response.data;
@@ -30,13 +39,42 @@ export class ListingEmployerComponent implements OnInit {
   }
 
   deleteEmployer(item: any){
-    console.log(item)
-    this.employerService.delete(`${DELETE_EMPLOYER}/${item.id}`)
-      .then((response:any)=>{
-      console.log('response', response)
-      this.getEmployer();
+    Swal.fire({
+      title: 'Êtes-vous sure?',
+      text: "Vous ne pourrez pas revenir en arrière !",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Oui, supprimez !'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        Swal.fire(
+          'Supprimé!',
+          'Employer supprimé.',
+          'success'
+        )
+        this.employerService.delete(`${DELETE_EMPLOYER}/${item.id}`)
+        .then((response:any)=>{
+        console.log('response', response)
+        this.getEmployer();
 
+      })
+      }
+      else {
+        Swal.fire(
+          'Annulé!',
+          'Employer non supprimé.',
+          'error'
+        )}
     })
   }
 
+  goTo(){
+    this.router.navigate(['/ajout-employer'])
+  }
+
+  recupId(employer: EmployerReponseModel){
+    this.router.navigate(['/ajout-employer/', employer.id])
+  }
 }
